add explicit children props for react 18, refactor LinkList
