+++ web/src/components/common/index.ts
@@ -2,6 +2,7 @@ export * from "./AlertPopup"
 export * from "./CollapsedAlert"
 export * from "./Dropdown"
 export * from "./LinkList"
+export * from "./LinkListItem"
 export * from "./LoadingButton"
 export * from "./LogoIcon"
 export * from "./NoData"
